requirements.txt
requests
numpy
rasterio
geopandas
python-dotenv
shapely
pyproj
scipy
